test: refactor integration tests

+++ src/test/integration/org/example/learningsystem/course/controller/CourseEnrollmentControllerIT.java
@@ -1,6 +1,6 @@
 package org.example.learningsystem.course.controller;
 
-import org.example.learningsystem.AbstractCommonIT;
+import org.example.learningsystem.common.AbstractCommonIT;
 import lombok.RequiredArgsConstructor;
 import org.example.learningsystem.course.model.Course;
 import org.example.learningsystem.course.service.CourseService;
@@ -15,19 +15,18 @@ import org.springframework.security.test.context.support.WithMockUser;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-import static org.example.learningsystem.util.CourseTestUtils.buildCourse;
-import static org.example.learningsystem.util.StudentTestUtils.buildStudent;
+import static java.util.Objects.nonNull;
+import static org.example.learningsystem.common.builder.CourseEnrollmentRequestBuilder.buildEnrollRequest;
+import static org.example.learningsystem.common.builder.CourseEnrollmentRequestBuilder.buildUnenrollRequest;
+import static org.example.learningsystem.common.util.CourseUtilsIT.buildCourse;
+import static org.example.learningsystem.common.util.StudentUtilsIT.buildStudent;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @Tag("integration")
 @RequiredArgsConstructor
 class CourseEnrollmentControllerIT extends AbstractCommonIT {
-
-    private static final String ENROLLMENT_URL = "/courses/{id}/students/{studentId}";
 
     @Autowired
     private CourseService courseService;
@@ -45,10 +44,12 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var student = buildStudentWithCoins(course.getPrice());
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
+
         var pageable = PageRequest.of(0, 1);
 
         // when
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isOk());
 
         // then
@@ -77,7 +78,8 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isInternalServerError());
 
         cleanUp(courseId, studentId);
@@ -96,7 +98,8 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isInternalServerError());
 
         cleanUp(courseId, studentId);
@@ -109,7 +112,8 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isUnauthorized());
     }
 
@@ -120,13 +124,15 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var course = buildPrivateCourse();
         var savedCourse = courseService.create(course);
         var courseId = savedCourse.getId();
+
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNotFound());
 
-        courseService.deleteById(courseId);
+        cleanUp(courseId, studentId);
     }
 
     @Test
@@ -134,15 +140,17 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
     void enrollStudent_givenCourseIdAndStudent_shouldThrowEntityNotFoundExceptionAndReturn404() throws Exception {
         // given
         var courseId = UUID.randomUUID();
+
         var student = buildStudentWithCoins(BigDecimal.ZERO);
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
 
         // when, then
-        mockMvc.perform(post(ENROLLMENT_URL, courseId, studentId))
+        var request = buildEnrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNotFound());
 
-        studentService.deleteById(studentId);
+        cleanUp(courseId, studentId);
     }
 
     @Test
@@ -156,10 +164,12 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var student = buildStudentWithCoins(course.getPrice());
         var savedStudent = studentService.create(student);
         var studentId = savedStudent.getId();
+
         var pageable = PageRequest.of(0, 10);
 
         // when
-        mockMvc.perform(delete(ENROLLMENT_URL, courseId, studentId))
+        var request = buildUnenrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isNoContent());
 
         // then
@@ -176,7 +186,8 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
         var studentId = UUID.randomUUID();
 
         // when, then
-        mockMvc.perform(delete(ENROLLMENT_URL, courseId, studentId))
+        var request = buildUnenrollRequest(courseId, studentId);
+        mockMvc.perform(request)
                 .andExpect(status().isUnauthorized());
     }
 
@@ -194,8 +205,12 @@ class CourseEnrollmentControllerIT extends AbstractCommonIT {
     }
 
     private void cleanUp(UUID courseId, UUID studentId) {
-        courseService.deleteById(courseId);
-        studentService.deleteById(studentId);
+        if (nonNull(courseId)) {
+            courseService.deleteById(courseId);
+        }
+        if (nonNull(studentId)) {
+            studentService.deleteById(studentId);
+        }
     }
 
 }
